Move dependency manager to a standalone mod
